requirements.txt
# Core dependencies
google-generativeai==0.3.0
pandas>=2.0.0
numpy>=1.24.0

# Utilities
python-dotenv>=1.0.0

# Recommended - faster processing
tqdm>=4.65.0
sentence-transformers>=2.2.0
scikit-learn>=1.0.0